[refactor] remove outdated files (replaced with command-based shell)
