--- app/run_list.lst
@@ -1,4 +1,0 @@
-TESTAPP1
-TESTAPP2
-TESTAPP3
-TESTAPP4
